ci: refactor test functions into their own test files
